+++ tests/test.c
@@ -633,6 +633,7 @@ struct Test tests[] = {
     TEST_CASE(test_crypto_crypto),
     TEST_CASE(test_crypto_aead),
     TEST_CASE(test_crypto_pbkdf2_hmac),
+    TEST_CASE(test_crypto_misc),
 
     // TEST CRASHES HERE: TEST_CASE(memlimit),